+++ amethyst/src/main/java/com/vitorpamplona/amethyst/model/nip51Lists/peopleList/FollowListsState.kt
@@ -20,291 +20,4 @@
  */
 package com.vitorpamplona.amethyst.model.nip51Lists.peopleList
 
-import com.vitorpamplona.amethyst.commons.model.anyNotNullEvent
-import com.vitorpamplona.amethyst.commons.model.eventIdSet
-import com.vitorpamplona.amethyst.commons.model.events
-import com.vitorpamplona.amethyst.commons.model.updateFlow
-import com.vitorpamplona.amethyst.model.Account
-import com.vitorpamplona.amethyst.model.AddressableNote
-import com.vitorpamplona.amethyst.model.LocalCache
-import com.vitorpamplona.amethyst.model.Note
-import com.vitorpamplona.amethyst.model.User
-import com.vitorpamplona.amethyst.model.filter
-import com.vitorpamplona.quartz.nip01Core.core.HexKey
-import com.vitorpamplona.quartz.nip01Core.signers.NostrSigner
-import com.vitorpamplona.quartz.nip01Core.signers.update
-import com.vitorpamplona.quartz.nip01Core.tags.dTag.dTag
-import com.vitorpamplona.quartz.nip09Deletions.DeletionEvent
-import com.vitorpamplona.quartz.nip51Lists.followList.FollowListEvent
-import com.vitorpamplona.quartz.nip51Lists.followList.description
-import com.vitorpamplona.quartz.nip51Lists.followList.image
-import com.vitorpamplona.quartz.nip51Lists.followList.person
-import com.vitorpamplona.quartz.nip51Lists.followList.personFirst
-import com.vitorpamplona.quartz.nip51Lists.followList.removePerson
-import com.vitorpamplona.quartz.nip51Lists.followList.title
-import com.vitorpamplona.quartz.nip51Lists.muteList.tags.UserTag
-import com.vitorpamplona.quartz.utils.flattenToSet
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.ExperimentalCoroutinesApi
-import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.SharingStarted
-import kotlinx.coroutines.flow.StateFlow
-import kotlinx.coroutines.flow.emitAll
-import kotlinx.coroutines.flow.flowOn
-import kotlinx.coroutines.flow.map
-import kotlinx.coroutines.flow.onStart
-import kotlinx.coroutines.flow.stateIn
-import kotlinx.coroutines.flow.transformLatest
-import kotlinx.coroutines.flow.update
-import java.util.UUID
-
-/**
- * Maintains several stateflows for each step in processing PeopleLists
- *
- * This class must receive updates from Large Cache as it receives new events.
- */
-class FollowListsState(
-    val signer: NostrSigner,
-    val cache: LocalCache,
-    val scope: CoroutineScope,
-) {
-    val user = cache.getOrCreateUser(signer.pubKey)
-
-    fun existingPeopleListNotes() = cache.addressables.filter(FollowListEvent.KIND, user.pubkeyHex)
-
-    val followListVersions = MutableStateFlow(0)
-
-    val followListNotes =
-        followListVersions
-            .map { existingPeopleListNotes() }
-            .onStart { emit(existingPeopleListNotes()) }
-            .flowOn(Dispatchers.IO)
-            .stateIn(scope, SharingStarted.Eagerly, emptyList())
-
-    val followListsEventIds =
-        followListNotes
-            .map { it.eventIdSet() }
-            .onStart { emit(followListNotes.value.eventIdSet()) }
-            .flowOn(Dispatchers.IO)
-            .stateIn(scope, SharingStarted.Eagerly, emptySet())
-
-    @OptIn(ExperimentalCoroutinesApi::class)
-    val latestLists: StateFlow<List<FollowListEvent>> =
-        followListNotes
-            .transformLatest { emitAll(it.updateFlow<FollowListEvent>()) }
-            .onStart { emit(followListNotes.value.events()) }
-            .flowOn(Dispatchers.IO)
-            .stateIn(scope, SharingStarted.Eagerly, emptyList())
-
-    fun List<FollowListEvent>.mapToUserIdSet() = this.map { it.followIdSet() }.flattenToSet()
-
-    val allPeopleListProfiles: StateFlow<Set<HexKey>> =
-        latestLists
-            .map { it.mapToUserIdSet() }
-            .onStart { emit(latestLists.value.mapToUserIdSet()) }
-            .flowOn(Dispatchers.IO)
-            .stateIn(scope, SharingStarted.Eagerly, emptySet())
-
-    fun FollowListEvent.toUI() =
-        PeopleList(
-            identifierTag = this.dTag(),
-            title = this.title() ?: this.dTag(),
-            description = this.description(),
-            image = this.image(),
-            privateMembers = emptySet(),
-            publicMembers = cache.load(this.followIdSet()),
-        )
-
-    fun List<FollowListEvent>.toUI() = this.map { it.toUI() }
-
-    val uiListFlow =
-        latestLists
-            .map { it.toUI() }
-            .onStart { emit(latestLists.value.toUI()) }
-            .flowOn(Dispatchers.IO)
-            .stateIn(scope, SharingStarted.Eagerly, emptyList())
-
-    fun List<PeopleList>.select(dTag: String) =
-        this.firstOrNull {
-            it.identifierTag == dTag
-        }
-
-    fun selectList(dTag: String) = uiListFlow.value.select(dTag)
-
-    fun selectListFlow(dTag: String) =
-        uiListFlow
-            .map { it.select(dTag) }
-            .onStart { emit(selectList(dTag)) }
-
-    fun isUserInFollowSets(user: User): Boolean = allPeopleListProfiles.value.contains(user.pubkeyHex)
-
-    fun DeletionEvent.hasDeletedAnyFollowList() = deleteAddressesWithKind(FollowListEvent.KIND) || deletesAnyEventIn(followListsEventIds.value)
-
-    fun hasItemInNoteList(notes: Set<Note>): Boolean =
-        notes.anyNotNullEvent { event ->
-            if (event.pubKey == signer.pubKey) {
-                event is FollowListEvent || (event is DeletionEvent && event.hasDeletedAnyFollowList())
-            } else {
-                false
-            }
-        }
-
-    fun newNotes(newNotes: Set<Note>) {
-        if (hasItemInNoteList(newNotes)) {
-            forceRefresh()
-        }
-    }
-
-    fun deletedNotes(deletedNotes: Set<Note>) {
-        if (hasItemInNoteList(deletedNotes)) {
-            forceRefresh()
-        }
-    }
-
-    fun forceRefresh() {
-        followListVersions.update { it + 1 }
-    }
-
-    // --------------
-    // Updating Lists
-    // --------------
-
-    fun getPeopleListNote(noteIdentifier: String): AddressableNote? = existingPeopleListNotes().find { it.dTag() == noteIdentifier }
-
-    fun getPeopleList(noteIdentifier: String): FollowListEvent = getPeopleListNote(noteIdentifier)?.event as FollowListEvent
-
-    fun User.toUserTag() = UserTag(this.pubkeyHex, this.bestRelayHint())
-
-    fun Set<User>.toUserTags() = map { it.toUserTag() }
-
-    suspend fun addFollowList(
-        name: String,
-        desc: String?,
-        image: String?,
-        member: User? = null,
-        isPrivate: Boolean = false,
-        account: Account,
-    ): String {
-        val dTag = UUID.randomUUID().toString()
-
-        val newListTemplate =
-            FollowListEvent.build(
-                name = name,
-                people = if (!isPrivate && member != null) listOf(member.toUserTag()) else emptyList(),
-                dTag = dTag,
-            ) {
-                if (desc != null) description(desc)
-                if (image != null) image(image)
-            }
-
-        val newList = signer.sign(newListTemplate)
-
-        account.sendMyPublicAndPrivateOutbox(newList)
-        return dTag
-    }
-
-    suspend fun updateMetadata(
-        name: String?,
-        desc: String?,
-        image: String?,
-        peopleList: PeopleList,
-        account: Account,
-    ) {
-        val listEvent = getPeopleList(peopleList.identifierTag)
-
-        val template =
-            listEvent.update {
-                if (name != null) title(name)
-                if (desc != null) description(desc)
-                if (image != null) image(image)
-            }
-
-        val newList = signer.sign(template)
-
-        account.sendMyPublicAndPrivateOutbox(newList)
-    }
-
-    suspend fun cloneFollowSet(
-        currentFollowPack: PeopleList,
-        customCloneName: String?,
-        customCloneDescription: String?,
-        account: Account,
-    ) {
-        val listEvent = getPeopleList(currentFollowPack.identifierTag)
-
-        val template =
-            listEvent.update {
-                // new list
-                dTag(UUID.randomUUID().toString())
-
-                // updates names
-                if (customCloneName != null) title(customCloneName)
-                if (customCloneDescription != null) description(customCloneDescription)
-            }
-
-        val newList = signer.sign(template)
-
-        account.sendMyPublicAndPrivateOutbox(newList)
-    }
-
-    suspend fun deleteFollowSet(
-        identifierTag: String,
-        account: Account,
-    ) {
-        val followListEvent = getPeopleList(identifierTag)
-        val deletionEvent = account.signer.sign(DeletionEvent.build(listOf(followListEvent)))
-        account.sendMyPublicAndPrivateOutbox(deletionEvent)
-    }
-
-    suspend fun addUserToSet(
-        user: User,
-        identifierTag: String,
-        account: Account,
-    ) {
-        val followListEvent = getPeopleList(identifierTag)
-
-        val template =
-            followListEvent.update {
-                person(user.pubkeyHex, user.bestRelayHint())
-            }
-
-        val newList = signer.sign(template)
-
-        account.sendMyPublicAndPrivateOutbox(newList)
-    }
-
-    suspend fun addUserFirstToSet(
-        user: User,
-        identifierTag: String,
-        account: Account,
-    ) {
-        val followListEvent = getPeopleList(identifierTag)
-
-        val template =
-            followListEvent.update {
-                personFirst(user.pubkeyHex, user.bestRelayHint())
-            }
-
-        val newList = signer.sign(template)
-
-        account.sendMyPublicAndPrivateOutbox(newList)
-    }
-
-    suspend fun removeUserFromSet(
-        user: User,
-        identifierTag: String,
-        account: Account,
-    ) {
-        val followListEvent = getPeopleList(identifierTag)
-
-        val template =
-            followListEvent.update {
-                removePerson(user.pubkeyHex)
-            }
-
-        val newList = signer.sign(template)
-
-        account.sendMyPublicAndPrivateOutbox(newList)
-    }
-}
+typealias FollowListsState = com.vitorpamplona.amethyst.commons.model.nip51Lists.peopleList.FollowListsState
