+++ amethyst/src/main/java/com/vitorpamplona/amethyst/model/topNavFeeds/FeedTopNavFilterState.kt
@@ -53,7 +53,7 @@ import kotlinx.coroutines.flow.transformLatest
 class FeedTopNavFilterState(
     val feedFilterListName: MutableStateFlow<TopFilter>,
     val kind3Follows: StateFlow<Kind3FollowListState.Kind3Follows>,
-    val allFollows: StateFlow<MergedFollowListsState.AllFollows>,
+    val allFollows: StateFlow<com.vitorpamplona.amethyst.commons.model.serverList.MergedFollowListsState.AllFollows>,
     val locationFlow: () -> StateFlow<LocationState.LocationResult>,
     val followsRelays: StateFlow<Set<NormalizedRelayUrl>>,
     val blockedRelays: StateFlow<Set<NormalizedRelayUrl>>,
